"[TeamCity] OEMx PantherLake_pantherlake_program_platform_Daily Daily Updating Version/ID: PantherLake_2493_00
Hsd-es-id: N/A"
Original commit date: Wed Dec 4 12:07:47 2024 -0800
Original commit hash: 6406b01615179e8baaca53e84b744b5b33ad3597

+++ OneSiliconPkg/IpBlock/MemoryInit/IncludePrivate/MrcOemPlatform.h
@@ -345,7 +345,7 @@ ReturnFromSmc (
 **/
 VOID
 SaDramReset (
-  IN UINT32 PciEBaseAddress,
+  IN UINT64 PciEBaseAddress,
   IN UINT32 ResetValue
   );
